Prep. 1.20.0 mute_update

+++ mute_update/Update_mute_exec2.sh
@@ -3,6 +3,8 @@
 # Update_mute_exec2.sh            #
 # mute (C)2023 kitamura_design    #
 
+echo " Updating [ mute ] ..."
+
 #### Copy Update source ####
   ### Move to HOME dir ###
   cd /var/tmp/mute_update 2>/dev/null
@@ -14,8 +16,8 @@
   fi
 
   ### Rewrite log files ###
+  echo "Rewrite log files..."
   sudo cp -RT /var/www/cgi-bin/log ./www/cgi-bin/log 2>/dev/null
-  #sudo cp -R /var/www/cgi-bin/log/ ./www/cgi-bin/log 2>/dev/null
 
   if [ $? != 0 ]; then
     echo "Location: /cgi-bin/Update/Update_mute_error.cgi"
@@ -24,6 +26,7 @@
   fi
 
   ### Rewrite mute.conf file ###
+  echo "Rewrite mute.conf files..."
   sudo cp /var/www/cgi-bin/etc/mute.conf ./www/cgi-bin/etc/mute.conf 2>/dev/null
 
   if [ $? != 0 ]; then
@@ -34,6 +37,7 @@
 
   ### Copy New source files ###
   #sudo chmod -R 777 /var/www
+  echo "Copy New source files..."
   sudo cp -RT ./www /var/www 2>/dev/null 1>/dev/null
 
   if [ $? != 0 ]; then
@@ -43,50 +47,59 @@
   fi
 
  	## Install Update-Check Service & Timer (1.06)
-	sudo mv /var/www/cgi-bin/Update/updchk.service /etc/systemd/system/updchk.service 2>/dev/null 1>/dev/null
+  if [ ! -e /etc/systemd/system/updchk.service ]; then
+	    sudo mv /var/www/cgi-bin/Update/updchk.service /etc/systemd/system/updchk.service 2>/dev/null 1>/dev/null
 
-   if [ $? != 0 ]; then
-    echo "Location: /cgi-bin/Update/Update_mute_error.cgi"
-    echo ''
-    exit 1
+      if [ $? != 0 ]; then
+        echo "Location: /cgi-bin/Update/Update_mute_error.cgi"
+        echo ''
+        exit 1
+      fi
+
+	    sudo mv /var/www/cgi-bin/Update/updchk.timer /etc/systemd/system/updchk.timer 2>/dev/null 1>/dev/null
+
+      if [ $? != 0 ]; then
+        echo "Location: /cgi-bin/Update/Update_mute_error.cgi"
+        echo ''
+        exit 1
+      fi
+
+	    sudo chmod 644 /etc/systemd/system/updchk.service /etc/systemd/system/updchk.timer
+
+      sudo systemctl daemon-reload
+	    sudo systemctl --now enable updchk.service 2>/dev/null 1>/dev/null
+	    sudo systemctl --now enable updchk.timer 2>/dev/null 1>/dev/null
   fi
-
-	sudo mv /var/www/cgi-bin/Update/updchk.timer /etc/systemd/system/updchk.timer 2>/dev/null 1>/dev/null
-
- if [ $? != 0 ]; then
-    echo "Location: /cgi-bin/Update/Update_mute_error.cgi"
-    echo ''
-    exit 1
-  fi
-
-	sudo chmod 644 /etc/systemd/system/updchk.service /etc/systemd/system/updchk.timer
-
-  sudo systemctl daemon-reload
-	sudo systemctl --now enable updchk.service 2>/dev/null 1>/dev/null
-	sudo systemctl --now enable updchk.timer 2>/dev/null 1>/dev/null
 
   ## Install Start-Up Sound Service (1.07b)
-	sudo mv /var/www/cgi-bin/Start/startUpSound.service /etc/systemd/system/startUpSound.service 2>/dev/null 1>/dev/null
+  if [ ! -e /etc/systemd/system/startUpSound.service ]; then
+    echo ""
+    echo "Install Start-Up Sound Service..."
+	  sudo mv /var/www/cgi-bin/Start/startUpSound.service /etc/systemd/system/startUpSound.service 2>/dev/null 1>/dev/null
   
-   if [ $? != 0 ]; then
-    echo "Location: /cgi-bin/Update/Update_mute_error.cgi"
-    echo ''
-    exit 1
+     if [ $? != 0 ]; then
+      echo "Location: /cgi-bin/Update/Update_mute_error.cgi"
+      echo ''
+      exit 1
+    fi
+
+    sudo chmod 644 /etc/systemd/system/startUpSound.service
+
+	  sudo systemctl daemon-reload
+	  sudo systemctl enable startUpSound.service 2>/dev/null 1>/dev/null
   fi
 
-  sudo chmod 644 /etc/systemd/system/startUpSound.service
-
-	sudo systemctl daemon-reload
-	sudo systemctl enable startUpSound.service 2>/dev/null 1>/dev/null
-
 	## Activate streaming (1.12.0)
-  CHK_streaming=$(sudo grep "server.stream-response-body = 1" /etc/lighttpd/lighttpd.conf)
+  CHK_streaming=$(sudo grep "server.stream-response-body = 1" /etc/lighttpd/lighttpd.conf || true)
 
   if [ -z "$CHK_streaming" ]; then 
+      echo " Activating Stream-response... "
 	    sudo sed -i -e '/server.port.*= */a server.stream-response-body = 1' /etc/lighttpd/lighttpd.conf
   fi
 
   ## Replace getcover (1.12.0)
+  echo ""
+  echo " Replacing getcover... "
 	sudo gcc -o getcover getcover.c
   sudo mv ./getcover /usr/local/bin/getcover
 
@@ -96,8 +109,79 @@
     exit 1
   fi
 
+  ## Install DLNA & AirPlay Renderers (1.20)
+  echo ''
+  echo ' Installing Media Renderers...'
+
+  ## DLNA Renderer (UpMPDcli) 
+  echo ""
+  echo " DLNA Renderer..."
+
+	# Get OS Code-name
+	OS_CODENAME=$(lsb_release -sc)
+	echo "Detected OS: ${OS_CODENAME}"
+
+	# Install UpMPDcli and Setting
+	if [ "${OS_CODENAME}" = "bullseye" ] || [ "${OS_CODENAME}" = "bookworm" ] || [ "${OS_CODENAME}" = "trixie" ]; then #Version Check
+		# Import repository key & Keylist
+	    sudo curl -L \
+	    https://www.lesbonscomptes.com/pages/lesbonscomptes.gpg \
+	    -o /usr/share/keyrings/lesbonscomptes.gpg
+
+	    sudo curl -L \
+ 	   "https://www.lesbonscomptes.com/upmpdcli/pages/upmpdcli-r${OS_CODENAME}.sources" \
+ 	   -o /etc/apt/sources.list.d/upmpdcli.sources
+
+		# Install upmpdcli
+	    sudo apt update
+	    sudo apt -o Acquire::Retries=3 install -y -q upmpdcli
+
+    	# Replace unit file
+    	if [ -f /usr/lib/systemd/system/upmpdcli.service ]; then
+    	    sudo rm /usr/lib/systemd/system/upmpdcli.service
+    	fi
+    	sudo cp ./upmpdcli.service /etc/systemd/system/upmpdcli.service
+		  sudo cp ./upmpdcli.conf /etc/upmpdcli.conf
+		  sudo cp ./mute_icon.png /usr/share/upmpdcli/mute_icon.png
+
+		  #Enable and Start
+      sudo systemctl daemon-reload
+      sudo systemctl enable --now upmpdcli
+	else
+    	echo " Skipping UpMPDcli: Unsupported OS version."
+	fi
+
+  ## Shairport-sync
+  echo ""
+  echo "AirPlay Reciever..."
+
+	# Install shairport-sync
+	sudo apt -o Acquire::Retries=3 install -y -q shairport-sync
+
+	# AirPlay (Shairport-sync) settings
+	sudo systemctl stop shairport-sync
+
+	# Remove old init script if exists
+	if [ -f /etc/init.d/shairport-sync ]; then
+	    sudo rm /etc/init.d/shairport-sync
+	fi
+
+	# Copy config and systemd service files from the same directory
+	sudo cp ./shairport-sync.conf /etc/shairport-sync.conf
+	sudo cp ./shairport-sync.service /etc/systemd/system/shairport-sync.service	
+
+	# Enable and Start
+	echo " Enabling and Starting AirPlay Reciever..."
+	sudo systemctl daemon-reload
+	sudo systemctl enable --now shairport-sync
+
+  echo ""
+  echo " Installing Media Renderers... Done"
+
 #### Finalize ####
-newVER=$(cat ./update.info | grep ver | cut -d "=" -f 2)
+#newVER=$(cat ./update.info | grep ver | cut -d "=" -f 2)
+newVER=$(grep '^ver=' ./update.info | cut -d '=' -f 2)
+
 sudo sed -i -e "s/ver.*/ver.${newVER}/" /etc/motd 2>/dev/null 1>/dev/null
 
   if [ $? != 0 ]; then
@@ -114,7 +198,8 @@ sudo sed -i -e "s/ver=.*/ver=${newVER}/" /var/www/cgi-bin/etc/mute.conf 2>/dev/n
     exit 1
   fi
 
-Dark_Mode=$(grep darkmode /var/www/cgi-bin/etc/mute.conf)
+Dark_Mode=$(grep darkmode /var/www/cgi-bin/etc/mute.conf || true)
+
 if [ "$Dark_Mode" = "darkmode=off" ]; then
   sudo cp /var/www/html/css/css_select/main_light.css /var/www/html/css/main.css
